+++ src/main/java/com/bokmcdok/butterflies/registries/CreativeTabRegistry.java
@@ -9,7 +9,6 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.RegistryObject;
 import org.jetbrains.annotations.NotNull;
 
-
 /**
  * This class registers new creative tabs for the Butterflies Mod
  */