+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### 6.6.0 (2025-08-27)
+- Add spawn eggs with textures for all custom mobs.
+
+### 6.5.6 (2025-08-27)
+- Stopped ALL bottled butterflies from glowing.
+
+### 6.5.5 (2025-08-26)
+- Renamed 'male' to 'variant' for consistency.
+- 
 ### 6.5.4
 - Removed the search bar from the butterfly creative mode tab.